Clean up `submit_test` folder in `CalcJob` dry-run tests (#2867)
The launcher tests running with `dry_run=True` will create the
`submit_test` folder in the current working directory that was not being
cleaned up. The test is adapted to delete this folder in the `tearDown`
of the unit test. Note that this is not the most elegant solution as it
depends on the `submit_test` name not changing, which is currently not
directly configurable and it not being another folder in the current
working directory. However, since it is not configurable and the cleanup
*has* to be done in a `tearDown`, because otherwise if the test fails
the cleanup might be skipped if placed in the test itself, there is
currently no better solution.