Capture `KeyboardInterrupt` in `Runner.run` and kill processes (#2744)
Currently, if a user runs a `Process` in a local interpreter and then
interrupts the interpreter the process will be lost. However, the node
will reflect the last active state. This can be confusing to new users,
who will still for example see a "process" in the `Waiting` state.

To prevent this situation, the `Runner._run` method is modified and
attaches listeners to the `SIGINT` and `SIGTERM` signals that when
emitted will trigger a clean function that will log the interrupt and
call `process.kill` on the process that is currently in the scope.
This should ensure that if a user runs a process in a local interpreter
and interrupts it, the process will be properly killed, which will cause
the node to properly reflect what has happened.

The current solution works for a single process being run, but has
issues for nested processes. Sometimes only the current process is
killed and the others in the chain are unaffected before the interpreter
shuts down.