Implement an isolated mode for the `Pluginable` sub of `click.Group` (#2878)
The `Pluginable` sub class of `click.Group` allows sub commands to be
added to it dynamically from external plugins through entry points.
However, this gave problems for the auto-docs generator in the
pre-commit script `utils/validate_consistency.py` which would include
these external commands, even though it shouldn't. To solve this, we
implement an plugin excluden mode for the `Pluginable` class which the
pre-commit script will enable before generating the documentation.