Respect option `-p/--profile` in `verdi profile show` (#2876)
Even when a non-default option was specified with `verdi profile show`
the default specified in the configuration was taken. The order of
precedence for profile in verdi commands should be:

 * Configuration default
 * Profile option `-p/--profile`
 * Command specific profile parameter

By setting an attribute `profile_option_used` in the `ctx.obj` of the
`verdi` command, other commands can see whether the user specified an
explicit profile that should potentially be respected.