unwrap future from controller recursively and handle killing future
Since `LoopCommunicator` is the communicator for all the rpc and
broadcast communication, process controlling operation via command line
such as `verdi process pause` which calling the
`controller.pause_process` will wrap the result in one more
`kiwipy.Future`. Therefore, the future get from controller need to be
unwrapped by `plumpy.unwrap_kiwi_future` first.

The future result return from `controller.kill_process` is the result of
rpc calling of `aiida.engine.processes.process::Process::kill` method,
which will return a future represent a boolean value to
indicate that the process and its child process is successfully killed.
So the list of futures is chained to a future and returned.